added admin page, deleting postings and comments of an user

+++ src/main/java/com/example/cf/manager/service/CommentService.java
@@ -2,6 +2,7 @@ package com.example.cf.manager.service;
 
 import com.example.cf.manager.domain.CommentInfo;
 import com.example.cf.manager.domain.PostingInfo;
+import com.example.cf.manager.domain.UserInfo;
 import com.example.cf.manager.dto.CommentInfoDto;
 import com.example.cf.manager.dto.PostingInfoDto;
 import com.example.cf.manager.repository.CommentRepository;
@@ -23,12 +24,9 @@ public class CommentService {
         return commentRepository.findById(id);
     }
 
-    public void delete(Long code){
-        commentRepository.deleteById(code);
-    }
-
-    public void findAndDeleteByPosting(PostingInfo posting){
-        commentRepository.deleteAllByPostinginfo(posting);
+    public void deleteComment(Long code, UserInfo userInfo){
+        if(findById(code).get().getUserinfo().getUserid().equals(userInfo.getUserid()))
+            commentRepository.deleteById(code);
     }
 
     public Long save(CommentInfoDto commentinfodto) {
